Move `AccountsViewModel` to `: feature:navigation:drawer`
