+++ feature/navigation/drawer/src/main/kotlin/app/k9mail/feature/navigation/drawer/NavigationDrawerModule.kt
@@ -1,5 +1,6 @@
 package app.k9mail.feature.navigation.drawer
 
+import app.k9mail.feature.navigation.drawer.legacy.AccountsViewModel
 import app.k9mail.feature.navigation.drawer.ui.DrawerViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.core.module.Module
@@ -8,4 +9,12 @@ import org.koin.dsl.module
 val navigationDrawerModule: Module = module {
 
     viewModel { DrawerViewModel() }
+
+    viewModel {
+        AccountsViewModel(
+            accountManager = get(),
+            messageCountsProvider = get(),
+            messageListRepository = get(),
+        )
+    }
 }
