+++ legacy/ui/legacy/src/main/java/com/fsck/k9/ui/account/KoinModule.kt
@@ -2,14 +2,9 @@ package com.fsck.k9.ui.account
 
 import app.k9mail.legacy.ui.account.AccountFallbackImageProvider
 import app.k9mail.legacy.ui.account.AccountImageLoader
-import app.k9mail.legacy.ui.account.AccountsViewModel
-import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
 val accountUiModule = module {
-    viewModel {
-        AccountsViewModel(accountManager = get(), messageCountsProvider = get(), messageListRepository = get())
-    }
     factory { AccountImageLoader(accountFallbackImageProvider = get()) }
     factory { AccountFallbackImageProvider(context = get()) }
     factory { AccountImageModelLoaderFactory(contactPhotoLoader = get(), accountFallbackImageProvider = get()) }
